requirements.txt
tensorflow==2.15.0
numpy==1.26.4
flask==3.0.3
Pillow==10.3.0
requests==2.31.0
streamlit==1.33.0